feat: folder restructuring
