+++ index.html
@@ -11,6 +11,6 @@
 
   <body>
     <div id="app"></div>
-    <script type="module" src="/src/main.ts"></script>
+    <script type="module" src="/src/demo/main.ts"></script>
   </body>
 </html>
