+++ package.json
@@ -103,6 +103,7 @@
     "stylelint-config-standard-vue": "^1.0.0",
     "typescript": "^4.5.4",
     "vite": "^2.9.9",
+    "vite-plugin-chunk-split": "^0.2.7",
     "vue-eslint-parser": "^9.0.3",
     "vue-tsc": "^0.35.0"
   },
