+++ src/components/BreedingRhombusSpinner.vue
