+++ src/components/FingerprintSpinner.vue
