+++ src/components/FulfillingSquareSpinner.vue
