+++ src/components/HalfCircleSpinner.vue
