+++ src/components/IntersectingCirclesSpinner.vue
