+++ src/components/LoopingRhombusesSpinner.vue
