+++ src/components/PixelSpinner.vue
@@ -5,7 +5,7 @@
 </template>
 
 <script>
-import utils from '../../services/utils'
+import utils from '../services/utils'
 
 export default {
   name: 'PixelSpinner',
