+++ src/components/ScalingSquaresSpinner.vue
