+++ src/components/SelfBuildingSquareSpinner.vue
