+++ src/components/SpringSpinner.vue
@@ -10,7 +10,7 @@
 </template>
 
 <script>
-import utils from '../../services/utils'
+import utils from '../services/utils'
 
 export default {
   name: 'SpringSpinner',
