+++ src/components/SwappingSquaresSpinner.vue
