+++ src/components/TrinityRingsSpinner.vue
