+++ src/demo/assets/logo.png
