+++ src/demo/components/SpinnerSamples.vue
@@ -137,26 +137,26 @@
 </template>
 
 <script lang="ts">
-import HollowDotsSpinner from './lib/HollowDotsSpinner.vue'
-import PixelSpinner from './lib/PixelSpinner.vue'
-import FlowerSpinner from './lib/FlowerSpinner.vue'
-import IntersectingCirclesSpinner from './lib/IntersectingCirclesSpinner.vue'
-import OrbitSpinner from './lib/OrbitSpinner.vue'
-import FingerprintSpinner from './lib/FingerprintSpinner.vue'
-import TrinityRingsSpinner from './lib/TrinityRingsSpinner.vue'
-import FulfillingSquareSpinner from './lib/FulfillingSquareSpinner.vue'
-import CirclesToRhombusesSpinner from './lib/CirclesToRhombusesSpinner.vue'
-import SemipolarSpinner from './lib/SemipolarSpinner.vue'
-import BreedingRhombusSpinner from './lib/BreedingRhombusSpinner.vue'
-import SwappingSquaresSpinner from './lib/SwappingSquaresSpinner.vue'
-import ScalingSquaresSpinner from './lib/ScalingSquaresSpinner.vue'
-import FulfillingBouncingCircleSpinner from './lib/FulfillingBouncingCircleSpinner.vue'
-import RadarSpinner from './lib/RadarSpinner.vue'
-import SelfBuildingSquareSpinner from './lib/SelfBuildingSquareSpinner.vue'
-import SpringSpinner from './lib/SpringSpinner.vue'
-import LoopingRhombusesSpinner from './lib/LoopingRhombusesSpinner.vue'
-import HalfCircleSpinner from './lib/HalfCircleSpinner.vue'
-import AtomSpinner from './lib/AtomSpinner.vue'
+import HollowDotsSpinner from '../../components/HollowDotsSpinner.vue'
+import PixelSpinner from '../../components/PixelSpinner.vue'
+import FlowerSpinner from '../../components/FlowerSpinner.vue'
+import IntersectingCirclesSpinner from '../../components/IntersectingCirclesSpinner.vue'
+import OrbitSpinner from '../../components/OrbitSpinner.vue'
+import FingerprintSpinner from '../../components/FingerprintSpinner.vue'
+import TrinityRingsSpinner from '../../components/TrinityRingsSpinner.vue'
+import FulfillingSquareSpinner from '../../components/FulfillingSquareSpinner.vue'
+import CirclesToRhombusesSpinner from '../../components/CirclesToRhombusesSpinner.vue'
+import SemipolarSpinner from '../../components/SemipolarSpinner.vue'
+import BreedingRhombusSpinner from '../../components/BreedingRhombusSpinner.vue'
+import SwappingSquaresSpinner from '../../components/SwappingSquaresSpinner.vue'
+import ScalingSquaresSpinner from '../../components/ScalingSquaresSpinner.vue'
+import FulfillingBouncingCircleSpinner from '../../components/FulfillingBouncingCircleSpinner.vue'
+import RadarSpinner from '../../components/RadarSpinner.vue'
+import SelfBuildingSquareSpinner from '../../components/SelfBuildingSquareSpinner.vue'
+import SpringSpinner from '../../components/SpringSpinner.vue'
+import LoopingRhombusesSpinner from '../../components/LoopingRhombusesSpinner.vue'
+import HalfCircleSpinner from '../../components/HalfCircleSpinner.vue'
+import AtomSpinner from '../../components/AtomSpinner.vue'
 
 export default {
   name: 'SpinnerSamples',
