+++ src/demo/main.ts
@@ -1,6 +1,4 @@
 import { createApp } from 'vue'
 import App from './App.vue'
 
-import './styles/index.scss'
-
 createApp(App).mount('#app')
