+++ src/lib.ts
@@ -1,23 +1,23 @@
-import HollowDotsSpinner from './components/lib/HollowDotsSpinner.vue'
-import PixelSpinner from './components/lib/PixelSpinner.vue'
-import FlowerSpinner from './components/lib/FlowerSpinner.vue'
-import IntersectingCirclesSpinner from './components/lib/IntersectingCirclesSpinner.vue'
-import OrbitSpinner from './components/lib/OrbitSpinner.vue'
-import FingerprintSpinner from './components/lib/FingerprintSpinner.vue'
-import TrinityRingsSpinner from './components/lib/TrinityRingsSpinner.vue'
-import FulfillingSquareSpinner from './components/lib/FulfillingSquareSpinner.vue'
-import CirclesToRhombusesSpinner from './components/lib/CirclesToRhombusesSpinner.vue'
-import SemipolarSpinner from './components/lib/SemipolarSpinner.vue'
-import BreedingRhombusSpinner from './components/lib/BreedingRhombusSpinner.vue'
-import SwappingSquaresSpinner from './components/lib/SwappingSquaresSpinner.vue'
-import ScalingSquaresSpinner from './components/lib/ScalingSquaresSpinner.vue'
-import FulfillingBouncingCircleSpinner from './components/lib/FulfillingBouncingCircleSpinner.vue'
-import RadarSpinner from './components/lib/RadarSpinner.vue'
-import SelfBuildingSquareSpinner from './components/lib/SelfBuildingSquareSpinner.vue'
-import SpringSpinner from './components/lib/SpringSpinner.vue'
-import LoopingRhombusesSpinner from './components/lib/LoopingRhombusesSpinner.vue'
-import HalfCircleSpinner from './components/lib/HalfCircleSpinner.vue'
-import AtomSpinner from './components/lib/AtomSpinner.vue'
+import HollowDotsSpinner from './components/HollowDotsSpinner.vue'
+import PixelSpinner from './components/PixelSpinner.vue'
+import FlowerSpinner from './components/FlowerSpinner.vue'
+import IntersectingCirclesSpinner from './components/IntersectingCirclesSpinner.vue'
+import OrbitSpinner from './components/OrbitSpinner.vue'
+import FingerprintSpinner from './components/FingerprintSpinner.vue'
+import TrinityRingsSpinner from './components/TrinityRingsSpinner.vue'
+import FulfillingSquareSpinner from './components/FulfillingSquareSpinner.vue'
+import CirclesToRhombusesSpinner from './components/CirclesToRhombusesSpinner.vue'
+import SemipolarSpinner from './components/SemipolarSpinner.vue'
+import BreedingRhombusSpinner from './components/BreedingRhombusSpinner.vue'
+import SwappingSquaresSpinner from './components/SwappingSquaresSpinner.vue'
+import ScalingSquaresSpinner from './components/ScalingSquaresSpinner.vue'
+import FulfillingBouncingCircleSpinner from './components/FulfillingBouncingCircleSpinner.vue'
+import RadarSpinner from './components/RadarSpinner.vue'
+import SelfBuildingSquareSpinner from './components/SelfBuildingSquareSpinner.vue'
+import SpringSpinner from './components/SpringSpinner.vue'
+import LoopingRhombusesSpinner from './components/LoopingRhombusesSpinner.vue'
+import HalfCircleSpinner from './components/HalfCircleSpinner.vue'
+import AtomSpinner from './components/AtomSpinner.vue'
 
 export {
   HollowDotsSpinner,
